+++ CHANGES.rst
@@ -7,6 +7,9 @@ Changelog
 - Latest mockup and `mf_config` updates.
   [petschki]
 
+- Customizable input widget table css class.
+  [petschki]
+
 
 3.0.0a1 (2022-06-28)
 --------------------